fix: properly declare redux store global
